build.sh: prevent shell from interpreting patterns.
We want zip to handle the include patterns.

Fixes the "bug" that *.json had to be included multiple times in the
list (first one was expanded to the top-level manifest.json file by the
shell, second one was also expanded to the individual messages.json
files within the _locales directory).