{content/preferences.html,js/{mailhops,preferences}.js}: prepare option for extra information.
The current code only shows the basic auth status and a description of
what it means. Some users (i.e., me) would like to see more information,
typically the whole header part that is responsible for the decision, so
prepare to show this information.

This commit does NOT hook it up just yet. This will be done in a later
commit.